+++ src/main/java/com/example/shopingplusassignment/domain/brand/dto/request/UpdateBrandRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.shopingplusassignment.domain.brand.dto;
+package com.example.shopingplusassignment.domain.brand.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -6,9 +6,9 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class CreateBrandRequestDto {
+public class UpdateBrandRequestDto {
 
-    @NotBlank(message = "등록할 브랜드 이름을 입력해주세요.")
+    @NotBlank(message = "수정할 브랜드 이름을 입력해주세요.")
     private final String brandName;
 
 }
